# Mimir query engine: add support for some annotations emitted by `rate()`