--- a/pkg/streamingpromql/functions.go
+++ b/pkg/streamingpromql/functions.go
@@ -70,7 +70,7 @@ func InstantVectorLabelManipulationFunctionOperatorFactory(name string, metadata
 //   - metadataFunc: The function for handling metadata
 //   - stepFunc: The function to handle a range vector step
 //   - needMetricNames: If true, stepFunc emits annotations that contain the metric names for each series
-func SingleRangeVectorFunctionOperatorFactory(name string, metadataFunc functions.SeriesMetadataFunction, stepFunc functions.RangeVectorStepFunction, needMetricNames bool) InstantVectorFunctionOperatorFactory {
+func SingleRangeVectorFunctionOperatorFactory(name string, metadataFunc functions.SeriesMetadataFunction, stepFunc functions.RangeVectorStepFunction, seriesValidationFunc functions.RangeVectorSeriesValidationFunction, needMetricNames bool) InstantVectorFunctionOperatorFactory {
 	return func(args []types.Operator, memoryConsumptionTracker *limiting.MemoryConsumptionTracker, annotations *annotations.Annotations, expressionPosition posrange.PositionRange) (types.InstantVectorOperator, error) {
 		if len(args) != 1 {
 			// Should be caught by the PromQL parser, but we check here for safety.
@@ -83,19 +83,19 @@ func SingleRangeVectorFunctionOperatorFactory(name string, metadataFunc function
 			return nil, fmt.Errorf("expected a range vector argument for %s, got %T", name, args[0])
 		}
 
-		return operators.NewFunctionOverRangeVector(inner, memoryConsumptionTracker, metadataFunc, stepFunc, needMetricNames, annotations, expressionPosition), nil
+		return operators.NewFunctionOverRangeVector(inner, memoryConsumptionTracker, metadataFunc, stepFunc, seriesValidationFunc, needMetricNames, annotations, expressionPosition), nil
 	}
 }
 
 // RangeVectorTransformationFunctionOperatorFactory creates an InstantVectorFunctionOperatorFactory for functions
-// that have exactly 1 argument (v range-vector), and drop the series __name__ label.
+// that have exactly 1 argument (v range-vector), drop the series __name__ label, and do no per-series validation.
 //
 // Parameters:
 //   - name: The name of the function
 //   - stepFunc: The function to handle a range vector step
 //   - needMetricNames: If true, stepFunc emits annotations that contain the metric names for each series
 func RangeVectorTransformationFunctionOperatorFactory(name string, stepFunc functions.RangeVectorStepFunction, needMetricNames bool) InstantVectorFunctionOperatorFactory {
-	return SingleRangeVectorFunctionOperatorFactory(name, functions.DropSeriesName, stepFunc, needMetricNames)
+	return SingleRangeVectorFunctionOperatorFactory(name, functions.DropSeriesName, stepFunc, nil, needMetricNames)
 }
 
 // These functions return an instant-vector.