--- a/pkg/streamingpromql/operators/function_over_range_vector.go
+++ b/pkg/streamingpromql/operators/function_over_range_vector.go
@@ -60,7 +60,7 @@ func NewFunctionOverRangeVector(
 		o.seriesValidationFunc = f.SeriesValidationFuncFactory()
 	}
 
-	if f.NeedsSeriesNamesForAnnotations || o.seriesValidationFunc != nil {
+	if f.NeedsSeriesNamesForAnnotations {
 		o.metricNames = &MetricNames{}
 	}