#### What this PR does

This PR builds on #8660, adding annotations for the following cases for `rate`:

* `rate` is called over a range with both floats and histograms
* `rate` is called over a series containing floats with a metric name that does not look like a counter
* `rate` is called over a range with gauge-type histograms

Performance is generally on par or better than `main`, with the exception of `rate` over native histograms which has regressed somewhat due to all the extra checks:

<details>
<summary>Benchmark results</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
                                                                                │   main.txt   │  step-3-without-lastatorbefore.txt  │
                                                                                │    sec/op    │    sec/op     vs base               │
Query/rate(a_1[1m]),_instant_query/Mimir-10                                       143.0µ ± 15%   144.9µ ±  4%        ~ (p=0.818 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps/Mimir-10                          159.6µ ±  3%   159.7µ ±  3%        ~ (p=0.937 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps/Mimir-10                         277.4µ ±  6%   245.6µ ±  1%  -11.49% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query/Mimir-10                                     444.7µ ±  4%   409.6µ ±  1%   -7.88% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps/Mimir-10                        1.292m ±  5%   1.176m ±  1%   -8.98% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps/Mimir-10                       8.002m ±  1%   7.541m ±  1%   -5.76% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query/Mimir-10                                    4.406m ±  1%   4.119m ±  3%   -6.50% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps/Mimir-10                       17.98m ±  2%   17.45m ±  1%   -2.92% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps/Mimir-10                      138.0m ±  1%   134.7m ±  2%   -2.34% (p=0.004 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps/Mimir-10                        1.121m ±  5%   1.072m ±  1%   -4.36% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps/Mimir-10                      71.25m ±  2%   69.46m ±  3%   -2.52% (p=0.015 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps/Mimir-10                      1.421 ±  9%    1.368 ±  2%   -3.67% (p=0.002 n=6)
Query/rate(nh_1[1m]),_instant_query/Mimir-10                                      140.2µ ±  7%   134.6µ ±  2%   -4.00% (p=0.002 n=6)
Query/rate(nh_1[1m]),_range_query_with_100_steps/Mimir-10                         203.6µ ±  0%   197.0µ ±  1%   -3.23% (p=0.002 n=6)
Query/rate(nh_1[1m]),_range_query_with_1000_steps/Mimir-10                        716.2µ ±  4%   696.3µ ±  2%   -2.79% (p=0.004 n=6)
Query/rate(nh_100[1m]),_instant_query/Mimir-10                                    583.0µ ±  3%   567.8µ ±  1%   -2.61% (p=0.002 n=6)
Query/rate(nh_100[1m]),_range_query_with_100_steps/Mimir-10                       5.419m ±  2%   5.283m ±  2%   -2.50% (p=0.009 n=6)
Query/rate(nh_100[1m]),_range_query_with_1000_steps/Mimir-10                      46.00m ±  0%   45.60m ±  0%   -0.88% (p=0.002 n=6)
Query/rate(nh_2000[1m]),_instant_query/Mimir-10                                   6.382m ±  0%   6.312m ±  0%   -1.09% (p=0.002 n=6)
Query/rate(nh_2000[1m]),_range_query_with_100_steps/Mimir-10                      94.88m ±  0%   94.87m ±  1%        ~ (p=0.818 n=6)
Query/rate(nh_2000[1m]),_range_query_with_1000_steps/Mimir-10                     873.7m ±  1%   880.6m ±  1%   +0.79% (p=0.015 n=6)
Query/rate(nh_1[1m]),_range_query_with_10000_steps/Mimir-10                       5.118m ±  1%   5.018m ±  3%   -1.95% (p=0.041 n=6)
Query/rate(nh_100[1m]),_range_query_with_10000_steps/Mimir-10                     433.5m ±  1%   437.6m ±  2%   +0.94% (p=0.002 n=6)
Query/rate(nh_2000[1m]),_range_query_with_10000_steps/Mimir-10                     9.112 ±  2%    9.155 ±  2%        ~ (p=0.310 n=6)
Query/rate(a_1[1d]),_instant_query/Mimir-10                                       819.9µ ±  9%   791.7µ ±  2%   -3.44% (p=0.004 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps/Mimir-10                          1.182m ±  5%   1.127m ±  2%   -4.68% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps/Mimir-10                         4.264m ±  2%   4.068m ±  1%   -4.60% (p=0.002 n=6)
Query/rate(a_100[1d]),_instant_query/Mimir-10                                     47.33m ±  1%   46.31m ±  1%   -2.15% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps/Mimir-10                        80.56m ±  1%   79.23m ±  1%   -1.66% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps/Mimir-10                       379.4m ±  1%   373.8m ±  1%   -1.46% (p=0.004 n=6)
Query/rate(a_2000[1d]),_instant_query/Mimir-10                                    842.0m ±  1%   832.4m ±  0%   -1.14% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps/Mimir-10                        1.529 ±  0%    1.507 ±  0%   -1.41% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps/Mimir-10                       7.445 ±  0%    7.398 ±  1%   -0.62% (p=0.009 n=6)
Query/rate(nh_1[1h]),_instant_query/Mimir-10                                      276.5µ ± 11%   266.8µ ±  2%   -3.52% (p=0.009 n=6)
Query/rate(nh_1[1h]),_range_query_with_100_steps/Mimir-10                         489.3µ ±  3%   490.1µ ±  2%        ~ (p=0.937 n=6)
Query/rate(nh_1[1h]),_range_query_with_1000_steps/Mimir-10                        2.221m ±  1%   2.324m ±  1%   +4.67% (p=0.002 n=6)
Query/rate(nh_100[1h]),_instant_query/Mimir-10                                    6.753m ±  1%   6.744m ± 11%        ~ (p=0.937 n=6)
Query/rate(nh_100[1h]),_range_query_with_100_steps/Mimir-10                       24.80m ±  1%   26.23m ±  1%   +5.79% (p=0.002 n=6)
Query/rate(nh_100[1h]),_range_query_with_1000_steps/Mimir-10                      190.0m ±  1%   202.8m ±  1%   +6.74% (p=0.002 n=6)
Query/rate(nh_2000[1h]),_instant_query/Mimir-10                                   122.1m ±  3%   122.2m ±  0%        ~ (p=0.818 n=6)
Query/rate(nh_2000[1h]),_range_query_with_100_steps/Mimir-10                      468.8m ±  2%   501.4m ±  1%   +6.97% (p=0.002 n=6)
Query/rate(nh_2000[1h]),_range_query_with_1000_steps/Mimir-10                      3.749 ±  1%    4.055 ±  1%   +8.17% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query/Mimir-10                       243.8µ ±  2%   243.0µ ±  1%        ~ (p=0.699 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps/Mimir-10          279.6µ ±  4%   271.6µ ±  2%        ~ (p=0.180 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps/Mimir-10         448.0µ ±  1%   429.4µ ±  2%   -4.17% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query/Mimir-10                   893.0µ ±  3%   866.2µ ±  4%   -2.99% (p=0.009 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps/Mimir-10      2.351m ±  2%   2.292m ±  1%   -2.51% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps/Mimir-10     15.30m ±  1%   14.95m ±  0%   -2.29% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query/Mimir-10                 9.955m ±  2%   9.752m ±  1%   -2.05% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps/Mimir-10    37.75m ±  3%   37.36m ±  1%   -1.05% (p=0.009 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps/Mimir-10   281.0m ±  1%   279.9m ±  1%   -0.39% (p=0.002 n=6)
geomean                                                                           12.06m         11.85m         -1.69%
```

</details>

Peak memory consumption is unchanged.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.